+++ src/main/java/be/geo_solutions/translate_api/repostitories/api/LanguageRepository.java
@@ -4,6 +4,5 @@ import be.geo_solutions.translate_api.core.model.Language;
 import be.geo_solutions.translate_api.core.gateways.LanguageGateway;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LanguageRepository extends CrudRepository<Language, Integer>, LanguageGateway {
-
+public interface LanguageRepository extends CrudRepository<Language, Long>, LanguageGateway {
 }
