+++ src/main/java/be/geo_solutions/translate_api/repostitories/api/TranslationRepository.java
@@ -4,6 +4,6 @@ import be.geo_solutions.translate_api.core.gateways.TranslationGateway;
 import be.geo_solutions.translate_api.core.model.Translation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TranslationRepository extends CrudRepository<Translation, Integer>, TranslationGateway {
+public interface TranslationRepository extends CrudRepository<Translation, Long>, TranslationGateway {
 
 }
